--- a/src/diff/dijkstra.rs
+++ b/src/diff/dijkstra.rs
@@ -5,7 +5,7 @@ use std::{cmp::Reverse, env};
 
 use crate::{
     diff::changes::ChangeMap,
-    diff::graph::{populate_change_map, set_neighbours, Edge, SeenMap, Vertex},
+    diff::graph::{get_neighbours, populate_change_map, Edge, SeenMap, Vertex},
     parse::syntax::Syntax,
 };
 use bumpalo::Bump;
@@ -21,7 +21,7 @@ fn shortest_vertex_path<'a, 'b>(
     vertex_arena: &'b Bump,
     size_hint: usize,
     graph_limit: usize,
-) -> Result<Vec<&'b Vertex<'a, 'b>>, ExceededGraphLimit> {
+) -> Result<Vec<(Edge, &'b Vertex<'a, 'b>)>, ExceededGraphLimit> {
     // We want to visit nodes with the shortest distance first, but
     // RadixHeapMap is a max-heap. Ensure nodes are wrapped with
     // Reverse to flip comparisons.
@@ -35,23 +35,22 @@ fn shortest_vertex_path<'a, 'b>(
     let end: &'b Vertex<'a, 'b> = loop {
         match heap.pop() {
             Some((Reverse(distance), current)) => {
+                if current.is_visited.get() {
+                    continue;
+                }
+                current.is_visited.set(true);
+
                 if current.is_end() {
                     break current;
                 }
-                if unsafe { (&*current.neighbours.get()).is_some() } {
-                    continue; // visited
-                }
-                set_neighbours(current, vertex_arena, &mut seen);
 
-                if let Some(neighbors) = unsafe { &*current.neighbours.get() } {
-                    for &(edge, next) in neighbors {
-                        let distance_to_next = distance + edge.cost();
+                for (edge, next) in get_neighbours(current, vertex_arena, &mut seen) {
+                    let distance_to_next = distance + edge.cost();
 
-                        if distance_to_next < next.shortest_distance.get() {
-                            next.shortest_distance.set(distance_to_next);
-                            next.predecessor.replace(Some(current));
-                            heap.push(Reverse(distance_to_next), next);
-                        }
+                    if distance_to_next < next.shortest_distance.get() {
+                        next.shortest_distance.set(distance_to_next);
+                        next.predecessor.set(Some((edge, current)));
+                        heap.push(Reverse(distance_to_next), next);
                     }
                 }
 
@@ -70,35 +69,15 @@ fn shortest_vertex_path<'a, 'b>(
         heap.len(),
     );
 
-    let mut current = Some(end);
-    let mut vertex_route: Vec<&'b Vertex<'a, 'b>> = vec![];
-    while let Some(node) = current {
-        vertex_route.push(node);
+    let mut current = end.predecessor.get();
+    let mut vertex_route = vec![];
+    while let Some((edge, node)) = current {
+        vertex_route.push((edge, node));
         current = node.predecessor.get();
     }
 
     vertex_route.reverse();
     Ok(vertex_route)
-}
-
-fn shortest_path_with_edges<'a, 'b>(
-    route: &[&'b Vertex<'a, 'b>],
-) -> Vec<(Edge, &'b Vertex<'a, 'b>)> {
-    let mut prev = route.first().expect("Expected non-empty route");
-
-    let mut cost = 0;
-    let mut res = vec![];
-
-    for vertex in route.iter().skip(1) {
-        let edge = edge_between(prev, vertex);
-        res.push((edge, *prev));
-        cost += edge.cost();
-
-        prev = vertex;
-    }
-    debug!("Found a path of {} with cost {}.", route.len(), cost);
-
-    res
 }
 
 /// Return the shortest route from the `start` to the end vertex.
@@ -112,39 +91,7 @@ fn shortest_path<'a, 'b>(
     graph_limit: usize,
 ) -> Result<Vec<(Edge, &'b Vertex<'a, 'b>)>, ExceededGraphLimit> {
     let start: &'b Vertex<'a, 'b> = vertex_arena.alloc(start);
-    let vertex_path = shortest_vertex_path(start, vertex_arena, size_hint, graph_limit)?;
-    Ok(shortest_path_with_edges(&vertex_path))
-}
-
-fn edge_between<'a, 'b>(before: &Vertex<'a, 'b>, after: &Vertex<'a, 'b>) -> Edge {
-    assert_ne!(before, after);
-
-    let mut shortest_edge: Option<Edge> = None;
-    if let Some(neighbours) = unsafe { &*before.neighbours.get() } {
-        for &(edge, next) in neighbours {
-            // If there are multiple edges that can take us to `next`,
-            // prefer the shortest.
-            if *next == *after {
-                let is_shorter = match shortest_edge {
-                    Some(prev_edge) => edge.cost() < prev_edge.cost(),
-                    None => true,
-                };
-
-                if is_shorter {
-                    shortest_edge = Some(edge);
-                }
-            }
-        }
-    }
-
-    if let Some(edge) = shortest_edge {
-        return edge;
-    }
-
-    panic!(
-        "Expected a route between the two vertices {:#?} and {:#?}",
-        before, after
-    );
+    shortest_vertex_path(start, vertex_arena, size_hint, graph_limit)
 }
 
 /// What is the total number of AST nodes?