--- a/src/diff/dijkstra.rs
+++ b/src/diff/dijkstra.rs
@@ -155,10 +155,10 @@ pub fn mark_syntax<'a>(
                 format!(
                     "{:20} {:20} --- {:3} {:?}",
                     x.1.lhs_syntax
-                        .get_ref()
+                        .get_side()
                         .map_or_else(|| "None".into(), Syntax::dbg_content),
                     x.1.rhs_syntax
-                        .get_ref()
+                        .get_side()
                         .map_or_else(|| "None".into(), Syntax::dbg_content),
                     x.0.cost(),
                     x.0,