--- a/src/diff/dijkstra.rs
+++ b/src/diff/dijkstra.rs
@@ -142,7 +142,7 @@ pub fn mark_syntax<'a>(
     let size_hint = lhs_node_count * rhs_node_count;
 
     let start = Vertex::new(lhs_syntax, rhs_syntax);
-    let vertex_arena = Bump::new();
+    let vertex_arena = Bump::with_capacity(size_hint);
 
     let route = shortest_path(start, &vertex_arena, size_hint, graph_limit)?;