… Cargo.lock
… Cargo.toml
… src/diff/mod.rs
… src/diff/sliders.rs
… src/diff/stack.rs
… src/diff/unchanged.rs
… src/parse/syntax.rs