@QuarticCat 2022-10-07
Further possible optimizations:

- Parallelize the `for ... in possibly_changed` loop in `main.rs`. Expect no high speedup on this benchmark since the number of vertices is extremely unbalanced.
- Parallelize the shortest path algo. Maybe you can utilize the fact that the algo is running on a DAG.
- Bidirectional Dijkstra's Algo.

All of them require a huge workload so I'm not gonna try.

This may be the last big performance improvement PR. I can no longer find a place to optimize.