@QuarticCat 2023-02-08
I found that I forgot to update the latest progress ([afe07ea](https://github.com/Wilfred/difftastic/pull/401/commits/afe07ea55faf02d8d4bed1f48cf7c71f389e6cb7) - [80eae29](https://github.com/Wilfred/difftastic/pull/401/commits/80eae29dc00d07d901c21cdecb97c67907302150)). They virtually have no effect on the speed, but the memory usage dropped down to 125MB, which is ~23% of the original.

To further optimize memory usage, you can:

- Remove unused fields from `SyntaxInfo`, including `previous_sibling`, `prev`, `num_after`, and `unique_id` (replaced by pointers as they are pinned). You might want to keep them for experimenting A* heuristics though. I haven't found a way that can easily remove & add these fields.
- Simplify `Syntax`, including
  - Change `Vec<T>` to `Box<[T]>`
  - Maybe concat `{open,close}_position` and `{open,close}_content`
  - Change `String` to `Box<str>` or `&str`
  - Remove `children`, as `head_child` + `next_sibling` is enough

These might give you another ~5MB reduction and that will probably result in a slightly better performance gained from better locality and better memory access pattern (`syntax.head_child` vs `syntax.children.get(0)`).