@QuarticCat 2023-02-08
I guess you are not willing to adopt some of my optimizations, which is fine. But there are still some harmless yet beneficial optimizations:

- Skip visited vertices and remove the `neighbor` field from `Vertex`
- Refactor parent stack (`pop_both_ancestor` + `pop_lhs_cnt` + `pop_rhs_cnt`)

BTW, I find a TODO in your current code:

> TODO: experiment with storing SyntaxId only, and have a HashMap from SyntaxId to &Syntax.

I don't think this is a good idea, even a Vec is better than HashMap here. Yet you still have a big problem: how to implement `Hash` for `Vertex`? Unless this `Vec` / `Hashmap` is a global / thread-local variable, you cannot access it inside `hash` impl. Storing references to it in vertices is even worse because you lose the memory advantage.

According to my profiling results, most of the time was spent on hash map operations. We know that hash function evenly distributes the key into different slots. In a large hashmap (larger than your cache), that means perfect cache misses. I have a [benchmark](https://github.com/QuarticCat/hashbrown-benchmark) to verify this idea and I hope it can bring you some inspiration.